requirements.txt
flask>=3.0.0
matplotlib>=3.2.2
numpy>=1.18.5
opencv-python>=4.1.1
requests>=2.23.0
scipy>=1.4.1
tensorflow>=2.10.0
tqdm>=4.64.0
Pillow>=9.2.0